+++ CHANGELOG.md
@@ -11,9 +11,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased][]
 
-- Remove `manage` from default features.
+[Unreleased]: https://github.com/trussed-dev/trussed-staging/compare/v0.3.0...HEAD
 
-[Unreleased]: https://github.com/trussed-dev/trussed-staging/compare/v0.2.0...HEAD
+-
+
+## [0.3.0][] - 2024-03-25
+
+[0.3.0]: https://github.com/trussed-dev/trussed-staging/compare/v0.2.0...v0.3.0
+
+- Remove `manage` from default features.
+- Implement `HkdfExtension` (moved from [Nitrokey/trussed-hkdf-backend][])
+
+[Nitrokey/trussed-hkdf-backend]: https://github.com/Nitrokey/trussed-hkdf-backend
 
 ## [0.2.0][] - 2024-03-15